# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "EnrolStep": {
        "description": "The account's first sign-in (or an accepted invitation): enrol the authenticator with this secret.",
        "properties": {
          "challenge": {
            "title": "Challenge",
            "type": "string"
          },
          "next": {
            "const": "enroll",
            "title": "Next",
            "type": "string"
          },
          "otpauth_uri": {
            "title": "Otpauth Uri",
            "type": "string"
          },
          "secret": {
            "title": "Secret",
            "type": "string"
          }
        },
        "required": [
          "next",
          "challenge",
          "secret",
          "otpauth_uri"
        ],
        "title": "EnrolStep",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IdpStart": {
        "properties": {
          "url": {
            "description": "the identity provider's hosted sign-in page; the agentledger_idp cookie binds the browser",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "IdpStart",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "MfaStep": {
        "description": "A second factor is always required: the account is enrolled, enter the current code.",
        "properties": {
          "challenge": {
            "title": "Challenge",
            "type": "string"
          },
          "next": {
            "const": "mfa",
            "title": "Next",
            "type": "string"
          }
        },
        "required": [
          "next",
          "challenge"
        ],
        "title": "MfaStep",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AgentLedger",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/idp/start": {
      "get": {
        "operationId": "idp_start",
        "parameters": [
          {
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "default": "login",
              "title": "Purpose",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "invite",
            "required": false,
            "schema": {
              "default": "",
              "title": "Invite",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdpStart"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Idp Start"
      }
    },
...
    "/api/auth/login": {
      "post": {
        "operationId": "auth_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/MfaStep"
                    },
                    {
                      "$ref": "#/components/schemas/EnrolStep"
                    }
                  ],
                  "title": "Response Auth Login"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Auth Login"
      }
    },
...
    "/api/coverage": {
      "get": {
        "description": "What the product actually supports, per form, year and jurisdiction (spec §7). Read by every signed-in user and\nby the deployment smoke check.",
        "operationId": "get_coverage",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Coverage",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Coverage"
      }
    },
...
    "/api/returns/individual": {
      "post": {
        "description": "Compute a Form 1040 from facts and source documents. Nothing is stored (so the deployment smoke check may\ncall it too).",
        "operationId": "compute_individual_return",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Compute Individual Return",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Individual Return"
      }
    },
...
  }
}